refactor Alert component and fix side bar navigate
